% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cell.R
\name{format_cell}
\alias{format_cell}
\title{Formattable Cell Editor}
\usage{
format_cell(
  font.family = "",
  font.weight = "normal",
  format.fun = "comma",
  digits = 2,
  ...
)
}
\arguments{
\item{font.family}{Font family of values within the selected cells.}

\item{font.weight}{Font size of values within the selected cells.}

\item{format.fun}{Function to format the values within the selected cells.}

\item{digits}{Number of digits to which the values within the selected cells will be rounded.}

\item{...}{Parameters passed to \code{style}.}
}
\value{
Formatted table.
}
\description{
Easily edit cells in a table using \code{formattable}.
}
