% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{Get Colors}
\usage{
get_colors(x, pos = "#8af3a3", neg = "#f49c9c", zero = "#ffffff", zero_val = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{pos}{Color of the value farthest above \code{zero_val}.}

\item{neg}{Color of the value farthest below \code{zero_val}.}

\item{zero}{Color assigned to \code{zero_val}.}

\item{zero_val}{The value you wish to represent 0 (the default, or median value in \code{x}).}
}
\value{
A data frame with the newly calculated columns. See \code{gradient} documentation for relevant information.
}
\description{
Create a matrix from vector to represent colors in gradient.
}
