% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_tile.R
\name{gradient_tile}
\alias{gradient_tile}
\title{Formattable Gradient Tile}
\usage{
gradient_tile(
  format.fun = "percent",
  format.digits = 2,
  font.family = "Arial",
  font.weight = "normal",
  border.radius = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{format.fun}{Function to format the values within the selected cells.}

\item{format.digits}{Number of digits to which the values within the selected cells will be rounded.}

\item{font.family}{Font family of values within the selected cells.}

\item{font.weight}{Font size of values within the selected cells.}

\item{border.radius}{Radius of the border.}

\item{color}{Color assigned to text.}

\item{...}{Parameters passed to \code{get_colors}.}
}
\value{
Formatted table.
}
\description{
More flexible alternative to \code{formattable::color_tile}.
}
